requirements.txt
streamlit
openai
google-search-results
requests
python-dotenv
